+++ src/Hacks/SafeMode.cpp
@@ -60,6 +60,7 @@ std::vector<std::string> hacks = {
     "show-hitboxes",
     "show-triggers",
     "coin-tracers",
+    "show-trajectory",
 };
 
 void Client::onPostSetup()
@@ -111,6 +112,8 @@ class $modify (PlayLayerExt, PlayLayer)
 
         if (hasHackedAttempt)
         {
+            return this->onQuit();
+
             if (m_isPracticeMode)
             {
                 this->showCompleteText();
@@ -158,6 +161,36 @@ class $modify(GJGameLevel) {
 
 class $modify(EndLevelLayerExt, EndLevelLayer)
 {
+    void playCoinEffect(float p0)
+    {
+        if (!hasHackedAttempt)
+            EndLevelLayer::playCoinEffect(p0);
+    }
+
+    void playCurrencyEffect(float p0)
+    {
+        if (!hasHackedAttempt)
+            EndLevelLayer::playCurrencyEffect(p0);
+    }
+
+    void playDiamondEffect(float p0)
+    {
+        if (!hasHackedAttempt)
+            EndLevelLayer::playDiamondEffect(p0);
+    }
+
+    void playEndEffect()
+    {
+        if (!hasHackedAttempt)
+            EndLevelLayer::playEndEffect();
+    }
+
+    void playStarEffect(float p0)
+    {
+        if (!hasHackedAttempt)
+            EndLevelLayer::playStarEffect(p0);
+    }
+
     void onInfo(CCObject*)
     {
         std::stringstream ss;